rework grammar no-spaces
