+++ test/test-grammar.txt
@@ -24,9 +24,9 @@ print 7/16 4/5 3/1
 print 7:8 3:2 1:5
 
 // Parse comments
-# not another comment
-$ also not a type of comment
-not a type of comment
+// # not another comment
+// $ also not a type of comment
+// not a type of comment
 // 1.0111213141E5 < commented number
 
 // valid strings start and end with: "" '' ``
