+++ test/tree/test-dev.json
@@ -194,10 +194,10 @@
             [
               "drive",
               [
-                9,
-                12,
-                7,
-                6
+                2,
+                2,
+                3,
+                3
               ]
             ],
             [
