+++ test/tree/test-errors.json
@@ -29,6 +29,11 @@
           4
         ]
       ],
+      "notes2": [
+        "notes",
+        "notes",
+        "notes"
+      ],
       "beat": [
         "euclid",
         8,
@@ -70,7 +75,11 @@
     "groups": {},
     "print": [
       "theArr",
-      "notes2",
+      [
+        "notes",
+        "notes",
+        "notes"
+      ],
       [
         "euclid",
         8,
@@ -156,6 +165,24 @@
               "@identifier": "theArr"
             }
           ]
+        }
+      },
+      {
+        "@list": {
+          "@name": "notes2",
+          "@params": {
+            "@function": {
+              "@name": "duplicate",
+              "@args": [
+                {
+                  "@identifier": "notes"
+                },
+                {
+                  "@number": 3
+                }
+              ]
+            }
+          }
         }
       },
       {
@@ -311,10 +338,9 @@
     "Error at line 14: Unexpected rArray: ']' at list arr myArr]<-",
     "Error at line 15",
     "Error at line 16: Unexpected rArray: ']' at list arr [myArr]]<-",
-    "Error at line 18",
+    "Error at line 18: Unexpected newObject: 'new' at list theArr [ 5 6 7 8 9 new<-",
     "Error at line 21: Unexpected identifier: 'kick' at newsample kick<-",
     "Error at line 23: Unexpected identifier: 'spread' at ring notes shuffle spread<-",
-    "Error at line 24: Unexpected lParam: '(' at ring notes2 duplicate\t(<-",
     "Error at line 30: Unexpected identifier: 'euclid' at ring beat2 join( euclid 8 3) euclid<-"
   ]
 }
