+++ test/tree/test-sample.json
@@ -37,7 +37,7 @@
       ]
     },
     "objects": {
-      "LD7Z0428": {
+      "3UPV058E": {
         "object": "sample",
         "type": [
           "kick_min",
@@ -73,7 +73,7 @@
           ],
           "add_fx": [],
           "name": [
-            "LD7Z0428"
+            "3UPV058E"
           ]
         }
       },
